add sassy prod build prompt

+++ prod.sh
@@ -10,27 +10,33 @@
 # Tutorial:
 # http://developer.android.com/tools/publishing/app-signing.html#signing-manually
 
-rm -rf www
-rm platforms/android/build/outputs/apk/expedition.apk
+read -p "Did you test a quest on the beta build? (y/N) " -n 1
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+  rm -rf www
+  rm platforms/android/build/outputs/apk/expedition.apk
 
-# Rebuild the web app files
-webpack --config ./webpack.dist.config.js
+  # Rebuild the web app files
+  webpack --config ./webpack.dist.config.js
 
-# iOS
-cordova build ios
+  # iOS
+  cordova build ios
 
-# Deploy web app to prod once apps built
-aws s3 cp www s3://app.expeditiongame.com --recursive
+  # Deploy web app to prod once apps built
+  aws s3 cp www s3://app.expeditiongame.com --recursive
 
-# Invalidate files on cloudfront
-aws cloudfront create-invalidation --distribution-id E24IJ45RK0D6J8 --invalidation-batch file://cloudfront-invalidations-prod.json
+  # Invalidate files on cloudfront
+  aws cloudfront create-invalidation --distribution-id E24IJ45RK0D6J8 --invalidation-batch file://cloudfront-invalidations-prod.json
 
-# Android: build the signed prod app
-# Last because it requires password entry
-cordova build --release android
-# Signing the release APK
-jarsigner -verbose -sigalg SHA1withRSA -digestalg SHA1 -keystore ../android-release-key.keystore platforms/android/build/outputs/apk/android-release-unsigned.apk expedition_android
-# Verification:
-jarsigner -verify -verbose -certs platforms/android/build/outputs/apk/android-release-unsigned.apk
-# Aligning memory blocks (takes less RAM on app)
-./zipalign -v 4 platforms/android/build/outputs/apk/android-release-unsigned.apk platforms/android/build/outputs/apk/expedition.apk
+  # Android: build the signed prod app
+  # Last because it requires password entry
+  cordova build --release android
+  # Signing the release APK
+  jarsigner -verbose -sigalg SHA1withRSA -digestalg SHA1 -keystore ../android-release-key.keystore platforms/android/build/outputs/apk/android-release-unsigned.apk expedition_android
+  # Verification:
+  jarsigner -verify -verbose -certs platforms/android/build/outputs/apk/android-release-unsigned.apk
+  # Aligning memory blocks (takes less RAM on app)
+  ./zipalign -v 4 platforms/android/build/outputs/apk/android-release-unsigned.apk platforms/android/build/outputs/apk/expedition.apk
+else
+  echo "Prod build cancelled until tested on beta."
+fi
